requirements.txt
torch==1.13.1
torchvision==0.14.1
matplotlib==3.7.1
numpy
google-cloud-aiplatform
kfp